When NGINX uses proxy pass, it can cache the results. But if the request has a cookie, it does not cache the results. It assumes that the user is logged in, and needs to pass through to the Cromlech server.
What this commit does, is ask if the user is logged in.  If the user is
not logged in, then the user is assumed to be anonymous, and no cookie
is returend.

That way the NGINX server can handle all of the anonymous requests, and the
site should scale much much better.